Move Home page and its styles to a separated folder.
