+++ front/src/pages/App/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Header from './Header/Header';
 import Footer from './Footer';
-import Home from '../Home';
+import Home from '../Home/Home';
 import AddItem from '../AddItem';
 import EditItem from '../EditItem';
 import NoMatch from '../NoMatch';
